fix np.int to int warnings

+++ src/tnsu/structure_matrix_constructor.py
@@ -49,15 +49,15 @@ def infinite_structure_matrix_dict(name: str):
     return structure_matrix[name]
 
 
-def square_peps_pbc(side: np.int):
+def square_peps_pbc(side: int):
     """
     Creates a structure matrix of a square lattice Tensor network with periodic boundary conditions (pbc)
     of shape (side, side). The total number of tensors in the network would be side^2.
     :param side: side-length of tensor network
     :return: a structure matrix
     """
-    n_tensors = np.int(np.square(side))
-    structure_matrix = np.zeros((n_tensors, 2 * n_tensors), dtype=np.int)
+    n_tensors = int(np.square(side))
+    structure_matrix = np.zeros((n_tensors, 2 * n_tensors), dtype=int)
     for i in range(n_tensors):
         structure_matrix[i, i] = 4
         structure_matrix[i, i + n_tensors] = 1
@@ -72,7 +72,7 @@ def square_peps_pbc(side: np.int):
     return structure_matrix
 
 
-def rectangular_peps_obc(height: np.int, width: np.int):
+def rectangular_peps_obc(height: int, width: int):
     """
     Creates a structure matrix of a rectangular lattice tensor network with open (non-periodic) boundary
     conditions (obc) of shape (height x width). The total number of tensors in the network would be height x width.
@@ -90,7 +90,7 @@ def rectangular_peps_obc(height: np.int, width: np.int):
                 edge_list.append((i, j, 4, i + 1, j, 2))
             if j < width - 1:
                 edge_list.append((i, j, 3, i, j + 1, 1))
-    structure_matrix = np.zeros(shape=[height * width, len(edge_list)], dtype=np.int)
+    structure_matrix = np.zeros(shape=[height * width, len(edge_list)], dtype=int)
 
     # fill in the structure matrix
     for edge_idx, edge in enumerate(edge_list):
